+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 -   `searchQuerystringParameter` to `FormElement` class
 -   `personalisation` and `externalIdGenerationOnSubmit` to `Form` class
+-   `FormPersonalisation` model class for `personalisation`
 
 ## 4.1.8 (2023-05-12)